+++ packages/kiarina-utils-common/src/kiarina/utils/config_registry/_types/config_alias.py
@@ -1,0 +1,4 @@
+from typing import TypeAlias
+
+ConfigAlias: TypeAlias = str
+"""A string that can be used as an alias for a ConfigName."""