+++ packages/kiarina-utils-common/src/kiarina/utils/config_registry/_types/config_name.py
@@ -1,0 +1,3 @@
+from typing import TypeAlias
+
+ConfigName: TypeAlias = str