+++ packages/kiarina-utils-common/src/kiarina/utils/config_registry/_views/resolved_config.py
@@ -1,0 +1,10 @@
+from typing import Generic, NamedTuple, TypeVar
+
+from .._types.config_name import ConfigName
+
+T = TypeVar("T")
+
+
+class ResolvedConfig(NamedTuple, Generic[T]):
+    name: ConfigName
+    config: T